pyproject.toml
[project]
name = "growthctl"
version = "0.1.2"
description = "Marketing as Code - Manage ad campaigns like infrastructure"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.21.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=14.0.0",
    "facebook-business>=24.0.0",
]

[project.scripts]
growthctl = "growthctl.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/growthctl"]

[tool.hatch.build.targets.sdist]
include = ["src/", "README.md", "pyproject.toml", "examples/"]
exclude = [".venv/", "docs/", "__pycache__/", "*.pyc", ".ruff_cache/"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
]
